+++ lib/ui/inbox/InboxHomePanel.dart
@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/semantics.dart';
+import 'package:illinois/ui/settings/SettingsNotificationsContentPanel.dart';
 import 'package:rokwire_plugin/model/auth2.dart';
 import 'package:rokwire_plugin/model/inbox.dart';
 import 'package:illinois/service/Analytics.dart';
@@ -14,7 +15,6 @@ import 'package:rokwire_plugin/service/localization.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
-import 'package:illinois/ui/settings/SettingsNotificationsPanel.dart';
 import 'package:illinois/ui/widgets/Filters.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
 import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
@@ -230,7 +230,7 @@ class _InboxHomePanelState extends State<InboxHomePanel> implements Notification
       visible: _showBanner,
       child:GestureDetector(
         onTap: (){
-          Navigator.of(context).push(CupertinoPageRoute(builder: (context) => SettingsNotificationsPanel()));
+          Navigator.of(context).push(CupertinoPageRoute(builder: (context) => SettingsNotificationsContentPanel(content: SettingsNotificationsContent.preferences)));
         },
         child:Container(
           padding: EdgeInsets.symmetric(vertical: 10, horizontal: 16),