+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Settings Content for Profile and Notifications. (Task in progress) [#1670](https://github.com/rokwire/illinois-app/issues/1670).
 ### Fixed
 - Fixed scrolling in Home panel after drag and drop operation [#1629](https://github.com/rokwire/illinois-app/issues/1629).
 ### Changed